+++ apps/sdk-sandbox/package.json
@@ -20,7 +20,7 @@
     "@delvtech/hyperdrive-js": "*",
     "@delvtech/hyperdrive-wasm": "*",
     "@uniswap/sdk-core": "^7.6.0",
-    "@uniswap/smart-order-router": "^4.18.1",
+    "@uniswap/smart-order-router": "^4.18.2",
     "@uniswap/v3-periphery": "^1.4.4",
     "ethers": "^5.7.2",
     "viem": "^2.22.19"
