+++ apps/sdk-sandbox/scripts/example.ts
@@ -1,14 +1,10 @@
 import { Drift } from "@delvtech/drift";
 import { viemAdapter } from "@delvtech/drift-viem";
-import { fixed } from "@delvtech/fixed-point-wasm";
 import { appConfig } from "@delvtech/hyperdrive-appconfig";
-import {
-  ReadHyperdrive,
-  ReadWriteHyperdrive,
-  zapAbi,
-} from "@delvtech/hyperdrive-js";
+import { ReadHyperdrive, ReadWriteHyperdrive } from "@delvtech/hyperdrive-js";
+import { CurrencyAmount, Percent, Token, TradeType } from "@uniswap/sdk-core";
+import { AlphaRouter, SwapType } from "@uniswap/smart-order-router";
 import { ethers } from "ethers";
-import { Address, encodePacked, maxInt256 } from "viem";
 import { publicClient, walletClient } from "../client";
 
 const zapsConfig = appConfig.zaps[707];
@@ -16,17 +12,6 @@ const drift = new Drift(viemAdapter({ publicClient, walletClient }));
 
 const poolAddress = "0x324395D5d835F84a02A75Aa26814f6fD22F25698";
 const earliestBlock = 20180617n;
-
-async function fetchSwapPath() {
-  const provider = new ethers.providers.JsonRpcProvider(process.env.RPC_URL);
-
-  // const router = new AlphaRouter({
-  //   chainId: 707,
-  //   provider,
-  // });
-
-  // console.log(router, "router");
-}
 
 // Write instance for transactions
 const writePool = new ReadWriteHyperdrive({
@@ -48,163 +33,53 @@ const poolContract = drift.contract({
   address: poolAddress,
 });
 
-// SAMPLE ASSET ID AND MATURITY
-const assetId: bigint =
-  452312848583266388373324160190187140051835877600158453279131187532666397056n;
-const maturity = 1755734400n;
+async function fetchSwapPath() {
+  const provider = new ethers.providers.JsonRpcProvider(process.env.RPC_URL);
+  const router = new AlphaRouter({
+    chainId: 1,
+    provider,
+  });
 
-async function executeZapOpenAndClose() {
-  try {
-    const account = walletClient?.account.address as Address;
-    console.log("[START] Account:", account);
+  // Define tokens
+  const DAI = new Token(
+    1,
+    "0x6B175474E89094C44Da98b954EedeAC495271d0F",
+    18,
+    "DAI",
+    "Dai Stablecoin"
+  );
+  const USDC = new Token(
+    1,
+    "0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48",
+    6,
+    "USDC",
+    "USD Coin"
+  );
+  // Swap amount: 1 DAI
+  const amountIn = ethers.utils.parseUnits("1", 18);
+  const currencyAmountIn = CurrencyAmount.fromRawAmount(
+    DAI,
+    amountIn.toString()
+  );
 
-    // Get initial open long details
-    const beforeDetails = await readPool.getOpenLongDetails({
-      account,
-      assetId,
-      options: { block: "latest" },
-    });
-    console.log("openLongDetails BEFORE openLong:", beforeDetails);
-
-    // Open Long
-    const { result, request } = await publicClient.simulateContract({
-      abi: writePool.contract.abi,
-      address: poolAddress,
-      chain: publicClient.chain,
-      functionName: "openLong",
-      account,
-      gas: 16125042n,
-      args: [
-        BigInt(2e18), // 2 base token
-        1n,
-        1n,
-        {
-          asBase: true,
-          destination: account,
-          extraData: "0x",
-        },
-      ],
-    });
-
-    const openTxHash = await walletClient?.writeContract({
-      ...request,
-    });
-    if (!openTxHash) throw new Error("No open transaction hash received");
-    console.log("Open Long tx hash:", openTxHash);
-
-    const txReceipt = await publicClient.waitForTransactionReceipt({
-      hash: openTxHash,
-    });
-    await drift.cache.clear();
-
-    if (!txReceipt) throw new Error("No open transaction receipt received");
-    if (txReceipt.status !== "success") {
-      console.error("Open Transaction failed:", txReceipt);
-      throw new Error("Open Transaction failed");
+  // Fetch the route for an exact input swap
+  const route = await router.route(
+    currencyAmountIn,
+    USDC,
+    TradeType.EXACT_INPUT,
+    {
+      recipient: "0xf39Fd6e51aad88F6F4ce6aB8827279cffFb92266", // replace with your address
+      slippageTolerance: new Percent(50, 10000), // 0.50% tolerance
+      deadline: Math.floor(Date.now() / 1000) + 1800, // 30 min deadline
+      type: SwapType.SWAP_ROUTER_02,
     }
-    console.log("Open tx receipt status:", txReceipt.status);
+  );
 
-    // Get open long details after opening
-    const afterDetails = await readPool.getOpenLongDetails({
-      account,
-      assetId,
-    });
-    console.log("openLongDetails AFTER openLong:", afterDetails);
-
-    // Approve zap (auxiliary) contract to manage the long position
-    const approvalReceipt = await writePool.contract.write("setApproval", {
-      amount: maxInt256,
-      tokenID: assetId,
-      operator: zapsConfig.address,
-    });
-    console.log("Approval tx hash:", approvalReceipt);
-
-    // Simulate closeLong to preview base amount out
-    const { result: previewBaseAmountOut } =
-      await publicClient.simulateContract({
-        abi: writePool.contract.abi,
-        address: poolAddress,
-        chain: publicClient.chain,
-        functionName: "closeLong",
-        account,
-        args: [
-          maturity,
-          BigInt(1e18), // 1 base token (example)
-          1n,
-          {
-            asBase: true,
-            destination: account,
-            extraData: "0x",
-          },
-        ],
-      });
-    // console.log("Preview base out (raw):", previewBaseAmountOut);
-    console.log(
-      "Preview base out (formatted):",
-      fixed(previewBaseAmountOut).format()
-    );
-
-    // Prepare zap swap parameters
-    const blockData = await publicClient.getBlock();
-    const deadline = blockData.timestamp + 60n;
-    const swapPath = encodePacked(
-      ["address", "uint24", "address"],
-      [
-        "0x6B175474E89094C44Da98b954EedeAC495271d0F", // DAI
-        100, // fee tier
-        "0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48", // USDC
-      ]
-    );
-
-    const swapTx = await walletClient
-      ?.writeContract({
-        abi: zapAbi,
-        chain: publicClient.chain,
-        address: zapsConfig.address,
-        functionName: "closeLongZap",
-        gas: 16125042n,
-        args: [
-          poolAddress,
-          maturity,
-          BigInt(1e18), // 1 base token
-          1n,
-          {
-            destination: zapsConfig.address,
-            asBase: true,
-            extraData: "0x",
-          },
-          {
-            amountIn: previewBaseAmountOut,
-            amountOutMinimum: 1n,
-            deadline,
-            path: swapPath,
-            recipient: account,
-          },
-          false,
-        ],
-      })
-      .catch((err) => {
-        console.error("closeLongZap failed:", err);
-        throw err;
-      });
-    if (!swapTx) throw new Error("No close position transaction hash received");
-    console.log("closeLongZap tx hash:", swapTx);
-
-    const receipt = await publicClient.waitForTransactionReceipt({
-      hash: swapTx,
-    });
-    await drift.cache.clear();
-
-    const openLongDetailsAfterZap = await readPool.getOpenLongDetails({
-      account,
-      assetId,
-    });
-    console.log("openLongDetails AFTER zap close:", openLongDetailsAfterZap);
-
-    return txReceipt;
-  } catch (error) {
-    console.error("Failed to execute zap open/close:", error);
-    throw error;
+  if (route) {
+    console.log("Swap path found:");
+    console.log(route.route[0]);
+  } else {
+    console.log("No swap path found.");
   }
 }
